# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api to update and create relations for items and more functions",
        "version": "0.1"
    },
    "paths": {
...
        "/get-especificacaobycliente": {
            "post": {
                "tags": [
                    "Relacional - Itens"
                ],
                "summary": "Get item specification by client and item",
                "operationId": "4ef2a5d420cb9f51c2342f038ca211b7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "client": {
                                        "description": "Client id",
                                        "type": "integer"
                                    },
                                    "item_estoque": {
                                        "description": "Item id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/create-especificacao": {
            "post": {
                "tags": [
                    "Relacional - Itens"
                ],
                "summary": "Create item specification",
                "operationId": "f1d32e5af26f95daab74a6dad45e1ffa",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "empresa": {
                                        "description": "Company",
                                        "type": "string"
                                    },
                                    "revenda": {
                                        "description": "Reseller",
                                        "type": "string"
                                    },
                                    "client": {
                                        "description": "Client id",
                                        "type": "integer"
                                    },
                                    "item_estoque": {
                                        "description": "Item id",
                                        "type": "integer"
                                    },
                                    "item_especificacao": {
                                        "description": "Item specification",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/createrelacao": {
            "post": {
                "summary": "Create item relationship",
                "operationId": "4a1f0ba0486f6474f4291a562de79557",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "empresa": {
                                        "description": "Company id",
                                        "type": "integer"
                                    },
                                    "revenda": {
                                        "description": "Reseller id",
                                        "type": "integer"
                                    },
                                    "item_estoque_principal": {
                                        "description": "Primary item id",
                                        "type": "integer"
                                    },
                                    "item_estoque_relacionado": {
                                        "description": "Related item id",
                                        "type": "integer"
                                    },
                                    "relacao_multipla": {
                                        "description": "Multiple relationship",
                                        "type": "string"
                                    },
                                    "client": {
                                        "description": "Client id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/getrelacaobycliente": {
            "post": {
                "summary": "Get relationship by client and primary item",
                "operationId": "ad2f298c331f2a19f58bc5916fcc07f0",
                "parameters": [
                    {
                        "name": "client",
                        "in": "query",
                        "description": "Client id",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "item_estoque_principal",
                        "in": "query",
                        "description": "Primary item id",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    }
}